+++ src/elements/ComportementCiblage.cpp
@@ -51,7 +51,7 @@ Personnage* ComportementCiblage::ciblerZone(std::vector<Personnage*> ciblesPossi
 			double ycible= ciblesPossibles[compteur]->getCoordonnees().getPosY();
 			double xvoisin= tousLesPersonnages[i]->getCoordonnees().getPosX();
 			double yvoisin= tousLesPersonnages[i]->getCoordonnees().getPosY();
-			if ((pow((xcible-xvoisin),2)+pow((ycible-yvoisin),2))< pow(range,2))
+			if ((pow((float)(xcible-xvoisin),2)+pow((float)(ycible-yvoisin),2))< pow((float)range,2))
 			{
 					nombredevoisins[compteur]++;
 			}
